Cleaup tests for threading
